requirements.txt
Flask==3.0.3
pandas==2.2.2
PyYAML==6.0.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
